Split target reference implementation from interface
